+++ test/integrations/rewriteframes.test.ts
@@ -4,7 +4,7 @@ import type { Event } from '@sentry/types';
 import { Platform } from 'react-native';
 
 import { createReactNativeRewriteFrames } from '../../src/js/integrations/rewriteframes';
-import { isExpo } from '../../src/js/utils/environment';
+import { isExpo, isHermesEnabled } from '../../src/js/utils/environment';
 import { mockFunction } from '../testutils';
 
 jest.mock('../../src/js/utils/environment');
@@ -29,6 +29,7 @@ describe('RewriteFrames', () => {
 
   beforeEach(() => {
     mockFunction(isExpo).mockReturnValue(false);
+    mockFunction(isHermesEnabled).mockReturnValue(false);
     jest.resetAllMocks();
   });
 
@@ -668,6 +669,8 @@ describe('RewriteFrames', () => {
   });
 
   it('should parse React Native errors on Android Hermes', async () => {
+    mockFunction(isHermesEnabled).mockReturnValue(true);
+
     const ANDROID_REACT_NATIVE_HERMES = {
       message: 'Error: lets throw!',
       name: 'Error',
@@ -714,28 +717,28 @@ describe('RewriteFrames', () => {
             filename: 'app:///index.android.bundle',
             function: 'value',
             lineno: 1,
-            colno: 31561,
+            colno: 31562,
             in_app: true,
           },
           {
             filename: 'app:///index.android.bundle',
             function: 'value',
             lineno: 1,
-            colno: 32776,
+            colno: 32777,
             in_app: true,
           },
           {
             filename: 'app:///index.android.bundle',
             function: 'anonymous',
             lineno: 1,
-            colno: 31603,
+            colno: 31604,
             in_app: true,
           },
           {
             filename: 'app:///index.android.bundle',
             function: 'value',
             lineno: 1,
-            colno: 33176,
+            colno: 33177,
             in_app: true,
           },
           {
@@ -747,42 +750,42 @@ describe('RewriteFrames', () => {
             filename: 'app:///index.android.bundle',
             function: 'receiveTouches',
             lineno: 1,
-            colno: 122512,
+            colno: 122513,
             in_app: true,
           },
           {
             filename: 'app:///index.android.bundle',
             function: 'Ue',
             lineno: 1,
-            colno: 77571,
+            colno: 77572,
             in_app: true,
           },
           {
             filename: 'app:///index.android.bundle',
             function: 'Ne',
             lineno: 1,
-            colno: 77238,
+            colno: 77239,
             in_app: true,
           },
           {
             filename: 'app:///index.android.bundle',
             function: '_e',
             lineno: 1,
-            colno: 127755,
+            colno: 127756,
             in_app: true,
           },
           {
             filename: 'app:///index.android.bundle',
             function: 'anonymous',
             lineno: 1,
-            colno: 77747,
+            colno: 77748,
             in_app: true,
           },
           {
             filename: 'app:///index.android.bundle',
             function: 'z',
             lineno: 1,
-            colno: 74642,
+            colno: 74643,
             in_app: true,
           },
           {
@@ -794,21 +797,21 @@ describe('RewriteFrames', () => {
             filename: 'app:///index.android.bundle',
             function: 'A',
             lineno: 1,
-            colno: 74709,
+            colno: 74710,
             in_app: true,
           },
           {
             filename: 'app:///index.android.bundle',
             function: 'N',
             lineno: 1,
-            colno: 74267,
+            colno: 74268,
             in_app: true,
           },
           {
             filename: 'app:///index.android.bundle',
             function: 'C',
             lineno: 1,
-            colno: 74126,
+            colno: 74127,
             in_app: true,
           },
           { filename: 'native', function: 'apply', in_app: true },
@@ -816,7 +819,7 @@ describe('RewriteFrames', () => {
             filename: 'app:///index.android.bundle',
             function: 'k',
             lineno: 1,
-            colno: 74094,
+            colno: 74095,
             in_app: true,
           },
           { filename: 'native', function: 'apply', in_app: true },
@@ -824,7 +827,7 @@ describe('RewriteFrames', () => {
             filename: 'app:///index.android.bundle',
             function: 'b',
             lineno: 1,
-            colno: 74037,
+            colno: 74038,
             in_app: true,
           },
           { filename: 'native', function: 'apply', in_app: true },
@@ -835,21 +838,21 @@ describe('RewriteFrames', () => {
             filename: 'app:///index.android.bundle',
             function: '_performSideEffectsForTransition',
             lineno: 1,
-            colno: 230843,
+            colno: 230844,
             in_app: true,
           },
           {
             filename: 'app:///index.android.bundle',
             function: 'anonymous',
             lineno: 1,
-            colno: 224280,
+            colno: 224281,
             in_app: true,
           },
           {
             filename: 'app:///index.android.bundle',
             function: 'onPress',
             lineno: 1,
-            colno: 452701,
+            colno: 452702,
             in_app: true,
           },
         ],